Fix snap service name. Eliminate unnecessary restarts, files.
This change fixes a few issues I found:

The systemd service name differs slightly from the constant in the
charm, which prevented the service from automatically starting.

interface:juju-info flags remain set during the lifetime of the
relation, so another flag is used to only refresh the service once on
relation join.

The files/ subdirectory does not seem to be used any longer so it's been
removed.